Move OpenCensusUtils to com.google.api.client.http to avoid circular dependency between packages
