+++ google-http-client/src/main/java/com/google/api/client/http/OpenCensusUtils.java
@@ -12,11 +12,9 @@
  * the License.
  */
 
-package com.google.api.client.util;
+package com.google.api.client.http;
 
-import com.google.api.client.http.HttpHeaders;
-import com.google.api.client.http.HttpRequest;
-import com.google.api.client.http.HttpStatusCodes;
+import com.google.api.client.util.Preconditions;
 import com.google.common.annotations.VisibleForTesting;
 
 import io.opencensus.contrib.http.util.HttpPropagationUtil;
