+++ google-http-client/src/test/java/com/google/api/client/http/OpenCensusUtilsTest.java
@@ -12,7 +12,7 @@
  * the License.
  */
 
-package com.google.api.client.util;
+package com.google.api.client.http;
 
 import com.google.api.client.http.HttpHeaders;
